requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
websockets>=12.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
